Adding patch to solve the getUtility / five.localsitemanager / Acquisition problem
Note: we discussed this issue a lot and everyone agrees this issue has nothing
to do with kss, it only manifested here sadly, do to changes that should
not have been irreversibly applied to the entire Plone code without
more cautions testing. We apply it here as a temporary workaround
because we need Plone and kss working - now.

svn path=/plone.app.kss/trunk/; revision=13789